requirements.txt
# Contenido de requirements.txt (¡Corregido!)
# Contenido de requirements.txt
fastapi
uvicorn[standard]
pydantic[email] 
pymongo
python-dotenv